Create weekdays.java
TODO:
Create a program that prompts the user for an integer that represents a day in 
the week with 1 representing Sunday, 2 representing Monday, and so on. Create a 
static getName method within the dayName class that takes an integer as a 
parameter and returns a String which contains the name of the day. Then, 
in the main method, print out the day name in the console. If the number entered
is not from 1-7, inform the user that the number they have entered in invalid.